+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/api/memories/AgileBoardIdMemory.kt
@@ -2,9 +2,14 @@ package com.atlassian.performance.tools.jirasoftwareactions.api.memories
 
 import com.atlassian.performance.tools.jiraactions.api.memories.Memory
 
-@Deprecated(message = "Use AgileBoardMemory",
+@Deprecated(
+    message = "Use Memory<Board>",
     replaceWith = ReplaceWith(
-        expression = "AgileBoardMemory",
-        imports = ["com.atlassian.performance.tools.jirasoftwareactions.api.memories.AgileBoardMemory"])
+        expression = "Memory<Board>",
+        imports = [
+            "com.atlassian.performance.tools.jiraactions.api.memories.Memory",
+            "com.atlassian.performance.tools.jirasoftwareactions.api.boards.Board"
+        ]
+    )
 )
 interface AgileBoardIdMemory : Memory<String>